requirements.txt
fastapi[standard]>=0.109.2
uvicorn[standard]>=0.27.1
pydantic>=2.6.1
pydantic-settings==2.1.0
python-dotenv>=1.0.1
openai>=1.12.0
numpy>=1.24.4
joblib>=1.3.0
pandas>=2.0.3
PyYAML>=6.0
